rename to tracker and add bedrock handler
